fix: split client files to use feature flags
