+++ src/components/shared/VercelToolbar.tsx
@@ -1,13 +1,7 @@
-"use client"
-import featureFlags from "@/feature_flags"
-import { FlagValues } from "@vercel/flags/react"
-import { VercelToolbar } from "@vercel/toolbar/next"
+import { features } from "@/components/shared/feature"
+import { DevToolbarClient } from "./VercelToolbarClient"
 
-export function DevToolbar() {
-  return (
-    <>
-      <FlagValues values={featureFlags} />
-      {process.env.NODE_ENV === "development" && <VercelToolbar />}
-    </>
-  )
+export async function DevToolbar() {
+  const featureFlags = await features()
+  return <DevToolbarClient featureFlags={featureFlags} />
 }
